// src/api/generated/model/companyUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyUpdate {
  name?: string | null;
  country_code?: string | null;
  industry_ids?: string[] | null;
  employees?: number | null;
  revenue_eur?: number | string | null;
  hq_city?: string | null;
  homepage_url?: string | null;
  careers_url?: string | null;
  newsroom_url?: string | null;
  linkedin_url?: string | null;
  notes?: string | null;
  tags?: string[] | null;
  is_tracked?: boolean | null;
}